add tsn file support (TileSplit Naming (file))
a tsn file contains the image path, whether it should be run verbosely, and the tile size, and then is followed by regular naming statements as would be in one with a .txt extension. this allows you to set the tilesplit.py file as the default application for tsn files and run tilesplit without using a terminal prompt. it is possible to do this with automator on macos, and may be possible on windows or linux, but i do not know that for certain. running tilesplit with a tsn file generates a log file in the same directory as the script is in.